# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Campus-Auth Web API",
    "version": "5.0.0-alpha.1",
    "description": "Baseline OpenAPI document for Campus-Auth. Generated as a static snapshot to enable `npm run typegen`. NOTE: handlers are not yet annotated with utoipa derive macros; this file is a manual baseline and should be regenerated from #[derive(utoipa::OpenApi)] once that integration lands. The hand-written frontend/src/api/types.ts remains the source of truth until then."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:50721"
    }
  ],
  "paths": {
...
    "/api/config": {
      "get": {
        "operationId": "get___api_config",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "500": {
            "description": "Internal Error",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "put": {
        "operationId": "put___api_config",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "500": {
            "description": "Internal Error",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        }
      },
      "patch": {
        "operationId": "patch___api_config",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "500": {
            "description": "Internal Error",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        }
      }
    },
...
    "/ws/logs": {
      "get": {
        "operationId": "get___ws_logs",
        "description": "WebSocket：日志实时推送 + 状态 watch 推送（双通道）",
        "responses": {
          "101": {
            "description": "Switching Protocols"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "500": {
            "description": "Internal Error",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  }
}